Finish react smart bundle

+++ runtimes/node/versions/latest/helpers/remix/bundle.sh
@@ -4,10 +4,13 @@ if [ -n "$OPEN_RUNTIMES_OUTPUT_DIRECTORY" ]; then
     cd $OPEN_RUNTIMES_OUTPUT_DIRECTORY
 fi
 
-mkdir -p .build
-mv ./* .build/
+ENTRYPOINT="./server/index.js"
+if [ -e "$ENTRYPOINT" ]; then
+    mkdir -p .build
+    mv ./* .build/
 
-mv .build/ build/
+    mv .build/ build/
 
-mv /usr/local/build/package*.json ./
-mv /usr/local/build/node_modules/ ./node_modules/
+    mv /usr/local/build/package*.json ./
+    mv /usr/local/build/node_modules/ ./node_modules/
+fi
